python-src/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
playwright>=1.44.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
google-cloud-tasks>=2.16.0
google-cloud-bigquery>=3.20.0
loguru>=0.7.2
python-dotenv>=1.0.1
pendulum>=3.0.0
pyproj>=3.6.0
shapely>=2.0.0
beautifulsoup4>=4.12.0
python-dateutil>=2.9.0
pandas>=2.2.2
openpyxl>=3.1.2
